atividades/api-gov/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
colorama==0.4.6
cryptography==46.0.7
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.1
h11==0.16.0
httpcore==1.0.9
idna==3.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
python-decouple==3.8
python-dotenv==1.2.2
requests==2.33.1
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
